+++ blocks/detail-viewer/src/components/StringRenderer/index.tsx
@@ -1,3 +1,4 @@
+import { useBlock } from '@appsemble/preact';
 import { Fragment, h, VNode } from 'preact';
 
 import type { RendererProps, StringField } from '../../../block';
@@ -5,11 +6,18 @@ import type { RendererProps, StringField } from '../../../block';
 /**
  * An element for a text type schema.
  */
-export default function StringRenderer({ field, value = '' }: RendererProps<StringField>): VNode {
+export default function StringRenderer({ data, field }: RendererProps<StringField>): VNode {
+  const { utils } = useBlock();
+
+  const label = utils.remap(field.label, data);
+  const value = utils.remap(field.name, data);
+
   return (
     <Fragment>
-      {field.label && <h6 className="title is-6">{field.label}</h6>}
-      <div className="content">{typeof value === 'string' ? value : JSON.stringify(value)}</div>
+      {label ? <h6 className="title is-6">{label}</h6> : null}
+      {value ? (
+        <div className="content">{typeof value === 'string' ? value : JSON.stringify(value)}</div>
+      ) : null}
     </Fragment>
   );
 }