+++ packages/preact-components/src/index.tsx
@@ -5,3 +5,4 @@ export { default as Loader } from './Loader';
 export { default as Location } from './Location';
 export { default as Message } from './Message';
 export { default as Select } from './Select';
+export { default as useObjectURL } from './useObjectURL';